requirements.txt
# JinniChirag MUA Chatbot — Python Dependencies
# Install: pip install -r requirements.txt

# ── Core framework ─────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9

# ── Database ───────────────────────────────────────────────────────────────
pymongo>=4.6.0

# ── Embeddings (local, CPU, ~90 MB) ───────────────────────────────────────
sentence-transformers>=2.5.0
numpy>=1.24.0

# ── NLU classifier ────────────────────────────────────────────────────────
scikit-learn>=1.4.0

# ── Emotion detection ─────────────────────────────────────────────────────
vaderSentiment>=3.3.2

# ── Fuzzy query correction ────────────────────────────────────────────────
rapidfuzz>=3.6.0